+++ rich/live_render.py
@@ -1,10 +1,13 @@
-from typing import Optional, Tuple
+from typing import Literal, Optional, Tuple
 
 from .console import Console, ConsoleOptions, RenderableType, RenderResult
 from .control import Control
 from .segment import Segment
+from .text import Text
 from .style import StyleType
 from ._loop import loop_last
+
+VerticalOverflowMethod = Literal["crop", "ellipsis", "visible"]
 
 
 class LiveRender:
@@ -15,9 +18,15 @@ class LiveRender:
         style (StyleType, optional): An optional style to apply to the renderable. Defaults to "".
     """
 
-    def __init__(self, renderable: RenderableType, style: StyleType = "") -> None:
+    def __init__(
+        self,
+        renderable: RenderableType,
+        style: StyleType = "",
+        vertical_overflow: VerticalOverflowMethod = "ellipsis",
+    ) -> None:
         self.renderable = renderable
         self.style = style
+        self.vertical_overflow = vertical_overflow
         self._shape: Optional[Tuple[int, int]] = None
 
     def set_renderable(self, renderable: RenderableType) -> None:
@@ -53,23 +62,30 @@ class LiveRender:
     def __rich_console__(
         self, console: Console, options: ConsoleOptions
     ) -> RenderResult:
+        _Segment = Segment
         style = console.get_style(self.style)
         lines = console.render_lines(self.renderable, options, style=style, pad=False)
-        _Segment = Segment
         shape = _Segment.get_shape(lines)
-        if self._shape is None:
-            self._shape = shape
-        else:
-            width1, height1 = shape
-            width2, height2 = self._shape
-            self._shape = (
-                max(width1, min(options.max_width, width2)),
-                max(height1, height2),
-            )
 
-        width, height = self._shape
-        lines = _Segment.set_shape(lines, width, height)
+        _, height = shape
+        if height > console.size.height:
+            if self.vertical_overflow == "crop":
+                lines = lines[: console.size.height]
+                shape = _Segment.get_shape(lines)
+            elif self.vertical_overflow == "ellipsis":
+                lines = lines[: (console.size.height - 1)]
+                overflow_text = Text(
+                    "...",
+                    overflow="crop",
+                    justify="center",
+                    end="",
+                    style="live.ellipsis",
+                )
+                lines.append(list(console.render(overflow_text)))
+                shape = _Segment.get_shape(lines)
+        self._shape = shape
+
         for last, line in loop_last(lines):
-            yield from _Segment.make_control(line)
+            yield from line
             if not last:
-                yield _Segment.line(is_control=True)
+                yield _Segment.line()