+++ rich/text.py
@@ -397,14 +397,14 @@ class Text(JupyterMixin):
         Returns:
             Style: A Style instance.
         """
+        # TODO: This is a little inefficient, it is only used by full justify
         if offset < 0:
             offset = len(self) + offset
-
         get_style = console.get_style
         style = get_style(self.style).copy()
         for start, end, span_style in self._spans:
-            if offset >= start and offset < end:
-                style += get_style(span_style)
+            if end > offset >= start:
+                style += get_style(span_style, default="")
         return style
 
     def highlight_regex(